+++ src/main/java/tk/cth451/transitrailmod/blocks/GlassFenceBlock.java
@@ -1,0 +1,53 @@
+package tk.cth451.transitrailmod.blocks;
+
+import java.util.List;
+
+import net.minecraft.block.Block;
+import net.minecraft.block.BlockFence;
+import net.minecraft.block.BlockFenceGate;
+import net.minecraft.block.material.Material;
+import net.minecraft.block.properties.PropertyBool;
+import net.minecraft.block.state.IBlockState;
+import net.minecraft.entity.Entity;
+import net.minecraft.entity.player.EntityPlayer;
+import net.minecraft.init.Blocks;
+import net.minecraft.util.AxisAlignedBB;
+import net.minecraft.util.BlockPos;
+import net.minecraft.util.EnumWorldBlockLayer;
+import net.minecraft.world.IBlockAccess;
+import net.minecraft.world.World;
+import tk.cth451.transitrailmod.TransitRailMod;
+import tk.cth451.transitrailmod.init.ModBlocks;
+
+public class GlassFenceBlock extends BlockFence {
+	
+	public GlassFenceBlock() {
+		super(Material.glass);
+		this.setUnlocalizedName("glass_fence");
+		this.setCreativeTab(TransitRailMod.tabTransitRail);
+	}
+	
+	@Override
+	public boolean canHarvestBlock(IBlockAccess world, BlockPos pos, EntityPlayer player) {
+		return true;
+	}
+	
+	@Override
+	public boolean isOpaqueCube() {
+		return false;
+	}
+	
+	@Override
+	public EnumWorldBlockLayer getBlockLayer() {
+		return EnumWorldBlockLayer.CUTOUT;
+	}
+	
+	@Override
+	public boolean canConnectTo(IBlockAccess worldIn, BlockPos pos) {
+		Block block = worldIn.getBlockState(pos).getBlock();
+		if (block == Blocks.glass || block == ModBlocks.turnstile_block) {
+			return true;
+		}
+		return super.canConnectTo(worldIn, pos);
+	}
+}